add search query
